Updated to show no Orders and No Products

+++ shoppingApp/screens/user/UserProductsScreen.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { FlatList, Platform, Button, Alert, } from 'react-native';
+import { FlatList, Platform, Button, Alert, View, Text, } from 'react-native';
 import { useSelector, useDispatch, } from 'react-redux';
 import ProductItem from '../../components/shop/ProductItem';
 import { HeaderButtons, Item, } from 'react-navigation-header-buttons';
@@ -16,7 +16,7 @@ const UserProductsScreen = ({ navigation, }) => {
   const deleteHandler = (id) => {
     Alert.alert('Are you sure you want to delete this item?', [
       { text: 'No', style: 'default' },
-      { 
+      {
         text: 'Yes',
         style: 'destructive',
         onPress: () => dispatch(productsActions.deleteProduct(id))
@@ -30,24 +30,32 @@ const UserProductsScreen = ({ navigation, }) => {
     });
   };
 
+  if (!userProducts.length) {
+    return (
+      <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center'}}>
+        <Text>No Products Found</Text>
+      </View>
+    )
+  }
+
   return (
-    <FlatList 
+    <FlatList
       data={userProducts}
       renderItem={(itemData) => (
-        <ProductItem 
+        <ProductItem
           title={itemData.item.title}
           price={itemData.item.price}
           imageUrl={itemData.item.imageUrl}
           onSelect={() => {editProductHandler(itemData.item.id)}}
         >
-          <Button 
+          <Button
             title="EDIT"
             onPress={() => editProductHandler(itemData.item.id)}
             color={Colors.primary}
           />
-          <Button 
-            title="DELETE" 
-            onPress={() => deleteHandler(itemData.item.id)} 
+          <Button
+            title="DELETE"
+            onPress={() => deleteHandler(itemData.item.id)}
             color={Colors.primary}
           />
         </ProductItem>
@@ -60,8 +68,8 @@ UserProductsScreen.navigationOptions = (navData) =>  ({
   headerTitle: 'Your Products',
   headerLeft: (
     <HeaderButtons HeaderButtonComponent={CustomHeaderButton}>
-      <Item 
-        title='Menu' 
+      <Item
+        title='Menu'
         iconName={Platform.OS !== 'ios' ?  'ios-menu' : 'md-menu'}
         onPress={()=> {navData.navigation.toggleDrawer()}}
       />
@@ -69,8 +77,8 @@ UserProductsScreen.navigationOptions = (navData) =>  ({
   ),
   headerRight: (
     <HeaderButtons HeaderButtonComponent={CustomHeaderButton}>
-      <Item 
-        title='Add' 
+      <Item
+        title='Add'
         iconName={Platform.OS !== 'ios' ?  'ios-create' : 'md-create'}
         onPress={()=> {navData.navigation.navigate('EditProducts')}}
       />
